POAP-20 DONE Move schemas to a single file
